pyproject.toml
[project]
name = "closedformfm"
version = "0.1.0"
description = "Code of the NeurIPS 2025 paper On the Closed-Form of Flow Matching: Generalization Does Not Arise from Target Stochasticity"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clean-fid>=0.1.35",
    "fld @ git+https://github.com/marcojira/fld.git",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "matplotlib>=3.10.7",
    "mlflow>=3.5.1",
    "pandas>=2.3.3",
    "pot>=0.9.6.post1",
    "setuptools>=80.9.0",
    "torchcfm>=1.0.7",
    "torchdyn>=1.0.6",
]

[tool.uv.sources]
fld = { git = "https://github.com/marcojira/fld.git" }
